Rename Builder to Renderer
